+++ actionmailbox/lib/action_mailbox/gem_version.rb
@@ -9,7 +9,7 @@ module ActionMailbox
   module VERSION
     MAJOR = 6
     MINOR = 1
-    TINY  = 0
+    TINY  = 1
     PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")