+++ actionmailbox/CHANGELOG.md
@@ -1,4 +1,5 @@
+Allow for ActionMailbox to use a different storage service rather than ActiveStorage's defaults.
 
-
+*Matt*
 
 Please check [6-1-stable](https://github.com/rails/rails/blob/6-1-stable/actionmailbox/CHANGELOG.md) for previous changes.